requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.7
openai==1.35.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2024.1
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
